Python Regex Example
The app.py script opens the input001.txt file. Based on the setup settings contained on the first line of the text file the script will do the following:

The first line of the text file includes the parameters for the analyzed column (3), a threshold (10), and threshold exceeded value (4).

Example:
When data in column 3 has a value higher than 10, for 4 consecutive times the timestamp for when this starts is printed out to the console.

Note: I am also printing out the entire column and timestamp data set as a cross-reference in the console. Only column 3 and 1 are currently functional.